AVX-512 lookup experiment
This commit contains an experiment in using AVX-512 SIMD intrinsics to
do the lookup. Because AVX-512 is not stable, it must be run with
nightly. In addition, it must be run with target-cpu=native on a CPU
with AVX-512.

The central technique is a perfect hash, but not minimal. The hash
generates unique values for each of the keys, in a 128-entry lookup
table. An important technique is the masked load, to ensure no values
past the end of the slice are loaded.

In spite of all this, it's 1.15 - 1.5x slower than the big prefix match
on my Zen 5 (AMD HX 370) machine. Because the code is basically
branchless, it may perform better on less predictable workloads.